:recycle: [refactor] 소켓 통신 코드 함수화
